backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.7
pydantic-settings==2.9.1
email-validator==2.2.0
python-dotenv==1.1.0
apscheduler==3.11.0
httpx==0.28.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pytest==8.3.5
ruff==0.11.8
black==25.1.0
